+++ test/StoreErrorHandlingTests.cs
@@ -3,6 +3,7 @@ using System;
 using NUnit.Framework;
 using Playdux.src.Store;
 using UniRx;
+using static Playdux.test.TestUtils.TestUtils;
 
 namespace Playdux.test
 {
@@ -10,28 +11,32 @@ namespace Playdux.test
     {
         private static SimpleTestState MethodThatThrows(SimpleTestState _) => throw new Exception();
 
+        private Store<SimpleTestState>? simpleStore;
+
+        [TearDown]
+        public void Teardown() { simpleStore?.Dispose(); }
+
         [Test]
         public void ErrorInSubscribeDoesNotBreakOtherObservers()
         {
-            SimpleTestState init = new SimpleTestState(42);
-            var store = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
+            SimpleTestState init = new(42);
+            simpleStore = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
 
             int notified = 0;
             int errorSeen = 0;
-            store.ObservableFor(state => state)
-                .Subscribe(
-                    onNext: _ => notified++,
-                    onError: _ => errorSeen++);
-            store.ObservableFor(state => state)
+            simpleStore.ObservableFor(state => state)
                 .Subscribe(
                     onNext: _ => throw new Exception(),
                     onError: _ => { });
+            simpleStore.ObservableFor(state => state)
+                .Subscribe(
+                    onNext: _ => notified++,
+                    onError: _ => errorSeen++);
 
-            try { store.Dispatch(new EmptyAction()); }
-            catch { /* ignored */ }
-            try { store.Dispatch(new EmptyAction()); }
-            catch { /* ignored */ }
+            simpleStore.Dispatch(new EmptyAction());
+            simpleStore.Dispatch(new EmptyAction());
 
+            BlockingWait();
             Assert.AreEqual(0, errorSeen, $"Saw {errorSeen} errors");
             Assert.AreEqual(2, notified, $"Saw {notified} notifications");
         }
@@ -39,25 +44,26 @@ namespace Playdux.test
         [Test]
         public void ErrorInStreamDoesNotBreakOtherObservers()
         {
-            SimpleTestState init = new SimpleTestState(42);
-            var store = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
+            SimpleTestState init = new(42);
+            simpleStore = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
 
             int notified = 0;
             int errorSeen = 0;
-            store.ObservableFor(state => state)
+            simpleStore.ObservableFor(state => state)
                 .Subscribe(
                     onNext: _ => notified++,
                     onError: _ => errorSeen++);
-            store.ObservableFor(state => state)
+            simpleStore.ObservableFor(state => state)
                 .Select(MethodThatThrows)
                 .Subscribe(
                     onNext: _ => { },
                     onError: _ => { }
                 );
 
-            store.Dispatch(new EmptyAction());
-            store.Dispatch(new EmptyAction());
+            simpleStore.Dispatch(new EmptyAction());
+            simpleStore.Dispatch(new EmptyAction());
 
+            BlockingWait();
             Assert.AreEqual(0, errorSeen, $"Saw {errorSeen} errors");
             Assert.AreEqual(2, notified, $"Saw {notified} notifications");
         }
@@ -65,26 +71,26 @@ namespace Playdux.test
         [Test]
         public void ErrorInSelectorDoesNotBreakOtherObservers()
         {
-            SimpleTestState init = new SimpleTestState(42);
-            var store = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
+            SimpleTestState init = new(42);
+            simpleStore = new Store<SimpleTestState>(init, TestReducers.IncrementNSimpleTestStateReducer);
 
-            int notified1 = 0;
-            int errors1 = 0;
-            store.ObservableFor(state => state)
+            int notified = 0;
+            int errors = 0;
+            simpleStore.ObservableFor(state => state)
                 .Subscribe(
-                    onNext: _ => notified1++,
-                    onError: _ => errors1++);
-            store.ObservableFor(TestSelectors.ErrorSimpleTestStateSelector)
+                    onNext: _ => notified++,
+                    onError: _ => errors++);
+            simpleStore.ObservableFor(TestSelectors.ErrorSimpleTestStateSelector)
                 .Subscribe(
                     onNext: _ => { },
-                    onError: _ => { }
-                );
+                    onError: _ => { });
 
-            store.Dispatch(new EmptyAction());
-            store.Dispatch(new EmptyAction());
+            simpleStore.Dispatch(new EmptyAction());
+            simpleStore.Dispatch(new EmptyAction());
 
-            Assert.AreEqual(0, errors1, $"1 Saw {errors1} errors");
-            Assert.AreEqual(2, notified1, $"1 Saw {notified1} notifications");
+            BlockingWait();
+            Assert.AreEqual(0, errors, $"Saw {errors} errors");
+            Assert.AreEqual(2, notified, $"Saw {notified} notifications");
         }
     }
 }